# web-client/openapi-schema.json
{"openapi":"3.1.0","info":{"title":"Titanic Passenger API","description":"Example REST API. Customize for your own entity.","version":"1.0.0"},"paths":{"/api/v1/passengers":{"get":{"tags":["passengers"],"summary":"List Passengers","description":"List passengers with optional filters.","operationId":"list_passengers_api_v1_passengers_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":2000,"minimum":1,"default":100,"title":"Limit"}},{"name":"pclass","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Pclass"}},{"name":"survived","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Survived"}},{"name":"sex","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sex"}},{"name":"embarked","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Embarked"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TitanicPassengerResponse"},"title":"Response List Passengers Api V1 Passengers Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["passengers"],"summary":"Create Passenger","description":"Create a passenger. Upserts if ticket already exists.","operationId":"create_passenger_api_v1_passengers_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TitanicPassengerCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TitanicPassengerResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/passengers/{ticket}":{"get":{"tags":["passengers"],"summary":"Get Passenger","description":"Get a passenger by ticket number.","operationId":"get_passenger_api_v1_passengers__ticket__get","parameters":[{"name":"ticket","in":"path","required":true,"schema":{"type":"string","title":"Ticket"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TitanicPassengerResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["passengers"],"summary":"Update Passenger","description":"Update a passenger by ticket number.","operationId":"update_passenger_api_v1_passengers__ticket__put","parameters":[{"name":"ticket","in":"path","required":true,"schema":{"type":"string","title":"Ticket"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TitanicPassengerCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TitanicPassengerResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["passengers"],"summary":"Delete Passenger","description":"Delete a passenger by ticket number.","operationId":"delete_passenger_api_v1_passengers__ticket__delete","parameters":[{"name":"ticket","in":"path","required":true,"schema":{"type":"string","title":"Ticket"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sse/echo":{"get":{"tags":["sse"],"summary":"Sse Echo","description":"Echo a message back as SSE events. For testing streaming integration.","operationId":"sse_echo_api_v1_sse_echo_get","parameters":[{"name":"message","in":"query","required":false,"schema":{"type":"string","default":"hello","title":"Message"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sse/playground/converse":{"post":{"tags":["sse"],"summary":"Playground Converse Stream","description":"Stream Bedrock Converse API response as SSE events.","operationId":"playground_converse_stream_api_v1_sse_playground_converse_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConverseRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sse/jobs/{job_id}":{"get":{"tags":["sse"],"summary":"Sse Job Status","description":"Stream job status updates via SSE. Closes on terminal status.","operationId":"sse_job_status_api_v1_sse_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jobs":{"post":{"tags":["jobs"],"summary":"Submit Job","description":"Submit a new background job.","operationId":"submit_job_api_v1_jobs_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobCreate"}}}},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["jobs"],"summary":"List Jobs","description":"List jobs with optional filters.","operationId":"list_jobs_api_v1_jobs_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":50,"title":"Limit"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"job_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Job Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/JobResponse"},"title":"Response List Jobs Api V1 Jobs Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/jobs/{job_id}":{"get":{"tags":["jobs"],"summary":"Get Job","description":"Get a job by ID.","operationId":"get_job_api_v1_jobs__job_id__get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"string","title":"Job Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","description":"Health check.","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/version":{"get":{"summary":"Version","description":"App version and build identifier.","operationId":"version_version_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ConverseRequest":{"properties":{"model_id":{"type":"string","title":"Model Id"},"messages":{"items":{"$ref":"#/components/schemas/ConverseTurn"},"type":"array","title":"Messages"},"system_prompt":{"type":"string","title":"System Prompt","default":""},"temperature":{"anyOf":[{"type":"number","maximum":1.0,"minimum":0.0},{"type":"null"}],"title":"Temperature","default":0.7},"max_tokens":{"type":"integer","maximum":4096.0,"minimum":1.0,"title":"Max Tokens","default":256},"top_p":{"anyOf":[{"type":"number","maximum":1.0,"minimum":0.0},{"type":"null"}],"title":"Top P"}},"type":"object","required":["model_id","messages"],"title":"ConverseRequest","description":"Request payload for the converse endpoint."},"ConverseTurn":{"properties":{"role":{"type":"string","title":"Role"},"text":{"type":"string","title":"Text"}},"type":"object","required":["role","text"],"title":"ConverseTurn","description":"A single turn in a conversation with role and text content."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobCreate":{"properties":{"job_type":{"type":"string","title":"Job Type","default":"passenger_analysis"},"input_data":{"additionalProperties":true,"type":"object","title":"Input Data"}},"type":"object","required":["input_data"],"title":"JobCreate","description":"Request body for submitting a new job."},"JobResponse":{"properties":{"job_id":{"type":"string","title":"Job Id"},"status":{"type":"string","title":"Status"},"job_type":{"type":"string","title":"Job Type"},"input_data":{"additionalProperties":true,"type":"object","title":"Input Data"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Metadata"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"},"created_at":{"type":"string","title":"Created At"},"updated_at":{"type":"string","title":"Updated At"}},"type":"object","required":["job_id","status","job_type","input_data","created_at","updated_at"],"title":"JobResponse","description":"Response body for a job record."},"TitanicPassengerCreate":{"properties":{"ticket":{"type":"string","title":"Ticket"},"name":{"type":"string","title":"Name"},"pclass":{"type":"integer","maximum":3.0,"minimum":1.0,"title":"Pclass"},"survived":{"type":"integer","maximum":1.0,"minimum":0.0,"title":"Survived"},"sex":{"type":"string","title":"Sex"},"age":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Age"},"sibsp":{"type":"integer","minimum":0.0,"title":"Sibsp","default":0},"parch":{"type":"integer","minimum":0.0,"title":"Parch","default":0},"fare":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Fare"},"cabin":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cabin"},"embarked":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Embarked"},"boat":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Boat"},"body":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Body"},"home_dest":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Home Dest"},"analysis":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Analysis"}},"type":"object","required":["ticket","name","pclass","survived","sex"],"title":"TitanicPassengerCreate","description":"Request body for creating/updating a passenger."},"TitanicPassengerResponse":{"properties":{"ticket":{"type":"string","title":"Ticket"},"name":{"type":"string","title":"Name"},"pclass":{"type":"integer","title":"Pclass"},"survived":{"type":"integer","title":"Survived"},"sex":{"type":"string","title":"Sex"},"age":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Age"},"sibsp":{"type":"integer","title":"Sibsp"},"parch":{"type":"integer","title":"Parch"},"fare":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Fare"},"cabin":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cabin"},"embarked":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Embarked"},"boat":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Boat"},"body":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Body"},"home_dest":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Home Dest"},"analysis":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Analysis"}},"type":"object","required":["ticket","name","pclass","survived","sex","sibsp","parch"],"title":"TitanicPassengerResponse","description":"Response body for a passenger record."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}